libs/utilities/requirements.txt
# Utilities Library Dependencies
# Core data processing
pandas>=1.5.0
numpy>=1.20.0

# HTTP requests
requests>=2.28.0

# Binary data parsing
pyarrow>=10.0.0

# Machine learning (for ai_utils and leeway modeling)
xgboost>=2.0.0
scikit-learn>=1.0.0
joblib>=1.3.0

# Statistical analysis (for race_utils)
statsmodels>=0.13.0

# Environment variables
python-dotenv>=0.19.0

# Date/time handling (for datetime_utils)
pytz>=2022.1
python-dateutil>=2.8.0

# InfluxDB client (for api_utils)
influxdb-client>=1.36.0
